removed GuiceContainerTest for now. ServiceLocator related tests will need to be moved to HK2LinkerTest. also moved test objects into separate package for clarity.
