+++ src/test/java/com/hubspot/dropwizard/guice/objects/ExplicitBindingService.java
@@ -1,4 +1,4 @@
-package com.hubspot.dropwizard.guice;
+package com.hubspot.dropwizard.guice.objects;
 
 public class ExplicitBindingService {
