+++ src/test/java/com/hubspot/dropwizard/guice/objects/InjectedBundle.java
@@ -1,4 +1,4 @@
-package com.hubspot.dropwizard.guice;
+package com.hubspot.dropwizard.guice.objects;
 
 import io.dropwizard.Bundle;
 import io.dropwizard.setup.Bootstrap;
