+++ src/test/java/com/hubspot/dropwizard/guice/objects/InjectedProvider.java
@@ -1,4 +1,4 @@
-package com.hubspot.dropwizard.guice;
+package com.hubspot.dropwizard.guice.objects;
 
 import javax.ws.rs.ext.Provider;
