+++ src/test/java/com/hubspot/dropwizard/guice/objects/InjectedResource.java
@@ -1,4 +1,4 @@
-package com.hubspot.dropwizard.guice;
+package com.hubspot.dropwizard.guice.objects;
 
 import javax.ws.rs.Path;
